+++ software/damper.h
@@ -15,9 +15,9 @@ class Damper {
     const Mass* GetM0();
     const Mass* GetM1();
     // Returns Mass Matrix
-    const bounded_matrix<float, 2, 2> GetM();
+    const bounded_matrix<double, 2, 2> GetM();
     // Returns Damping Matrix
-    const bounded_matrix<float, 2, 2> GetC();
+    const bounded_matrix<double, 2, 2> GetC();
 
     double Get_c() { return this->c; };
